ci: reproducible builds
I'm lazy and I'm always recreating the repo from scratch instead of
updating changed packages only. This results in different timestamps in
packages and therefore also different signatures.

When syncing the repository from clients they can have several packages
(and signatures) in their cache. Reinstalling a package that is also in
cache results in Pacman complaining about invalid signature.
I suspect that is from the signature mismatch between the new version of
package in the repo and in the cache.